% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or95ci.R
\name{OR_95CI}
\alias{OR_95CI}
\title{Title Creates a confidence interval for the Odds Ratio (OR)}
\usage{
OR_95CI(coef, se, siglevel, roundto)
}
\arguments{
\item{coef}{The regression coefficient (Numerical)}

\item{se}{The standard error of the regression coefficient (Numerical)}

\item{siglevel}{The desired signficane level (Numerical)}

\item{roundto}{How many decimal points to round to (Numerical)}
}
\value{
A Confidence Interval OR (LB, UB)
}
\description{
Returns a confidence interval for the odds ratio for the desired significance level.
}
\examples{
OR_95CI(2, 0.5, 0.05, 2)
load(file = "data/toydata.rda")
mod <- glm(y~x1+x2, family = "binomial", data = toydata)
summ <- summary(mod)
coef <- summ$coefficients[2,1] # coefficient for x1
se <- summ$coefficients[2,2] # se for x1
OR_95CI(coef, se, 0.05, 2)
}
\author{
Ente Kang
}
