% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Example data}
\format{
A data frame with 100 rows and 3 variables:
\describe{
  \item{y}{The response variable, binary (0,1)}
  \item{x1}{The first explanatory variable, continuous}
  \item{x2}{The second explanatory variable, binary (0,1)}
}
}
\source{
https://q.utoronto.ca/courses/317202/files/28332455?module_item_id=5163390
}
\usage{
toydata
}
\description{
A toy dataset
}
\keyword{datasets}
